Added cannons to ships
An object of type Spaceship now has a two dimensional array called cannons. The first dimension of it tells you how many cannons a ship has and the second one specifies their x and y coordinates on the ship.
The aim of this decision was to be able to tell a ship from which cannon it should shoot, because half of the EnemyShips have more than one.
All cannons have now been added to all seven ships in the game and the Spaceship.shoot() method has been changed accordingly.
Additionaly, all projectiles now have a distinct color, depending on which type of ship they were shot by.